brain/requirements.txt
# Brain module third-party dependencies
flask>=3.0,<4
requests>=2.31,<3
jsonschema>=4.20,<5
pydantic>=2.5,<3
PyYAML>=6.0,<7
werkzeug>=3.0,<4
pytest-timeout>=2.2,<3
google-genai>=1.0,<2
chromadb>=0.5,<1
langchain-chroma>=0.2,<1
langchain-core>=0.3,<1
langchain-openai>=0.3,<1
langchain-text-splitters>=0.3,<1
tiktoken>=0.8,<1
